👌 IMPROVE: Added ItemCart component
