+++ src/components/pages/Home.js
@@ -6,6 +6,7 @@ import { Link } from 'react-router-dom';
 import { mixins } from '../../styles';
 import SearchBar from '../SearchBar';
 import Pagination from '../Pagination';
+import ItemCard from './../ItemCard';
 
 const MainBody = styled.div``;
 
@@ -20,61 +21,6 @@ const Loading = styled.div`
 	text-align: center;
 	padding-top: 4rem;
 	font-size: 2rem;
-`;
-
-const ProductContainer = styled.div`
-	display: flex;
-	flex-direction: column;
-	text-align: left;
-	border: 1px solid rgba(0, 0, 0, 0.3);
-	padding: 1rem;
-`;
-
-const Image = styled.img`
-	width: 300px;
-	height: 300px;
-	padding: 1rem;
-`;
-
-const Name = styled.div`
-	font-size: 1.2rem;
-	margin: 0.3rem 0.5rem;
-`;
-
-const Rating = styled.div`
-	display: flex;
-	align-content: center;
-	text-align: center;
-	margin: 0.3rem 0.5rem;
-`;
-
-const RatingNum = styled.div`
-	margin: 0.5rem;
-	font-size: 1.1rem;
-`;
-
-const Price = styled.div`
-	margin: 0.3rem 0.5rem;
-	font-size: 1.3rem;
-	font-weight: bold;
-`;
-
-const OnSale = styled.span`
-	height: 1rem;
-	width: 2rem;
-	background: #d22b2b;
-	color: white;
-	padding: 0.2rem 0.2rem;
-	margin-left: 0.5rem;
-	border-radius: 3px;
-`;
-
-const Wrapper = styled.div`
-	${mixins.flexCenter};
-`;
-
-const StyledLink = styled(Link)`
-	${mixins.yellowButton};
 `;
 
 function Home() {
@@ -97,8 +43,8 @@ function Home() {
 	);
 
 	const paginate = (number) => {
-		setCurrentPage(number)
-	}
+		setCurrentPage(number);
+	};
 
 	const handleOnSearch = (searchQuery) => {
 		fetch(URL)
@@ -117,8 +63,8 @@ function Home() {
 	};
 
 	useEffect(() => {
-		window.scrollTo(0, 0)
-	  }, [currentPage])
+		window.scrollTo(0, 0);
+	}, [currentPage]);
 
 	return (
 		<>
@@ -128,31 +74,16 @@ function Home() {
 				<MainBody>
 					<SearchBar searchError={error} onSearch={handleOnSearch} />
 					<MainContainer>
-						{products &&
-							currentProducts.map(
-								({ imageUrl, name, avgRating, price, isOnSale, _id }) => (
-									<ProductContainer key={_id}>
-										<Image src={imageUrl} alt={name}></Image>
-										<Name>{name}</Name>
-										<Rating>
-											<Stars rating={avgRating} id={_id}></Stars>
-											<RatingNum>{avgRating}</RatingNum>
-										</Rating>
-										<Price>
-											${price}
-											{isOnSale === true ? <OnSale>On Sale</OnSale> : ''}
-										</Price>
-										<Wrapper>
-											<StyledLink to={`/item/${_id}`}>View Item</StyledLink>
-										</Wrapper>
-									</ProductContainer>
-								)
-							)}
+						{products && currentProducts.map((data) => <ItemCard {...data} />)}
 					</MainContainer>
 					{error ? (
 						<h1 style={{ textAlign: 'center' }}>No results to display</h1>
 					) : (
-						<Pagination productsPerPage={productsPerPage} totalProducts={products.length} paginate={paginate}/>
+						<Pagination
+							productsPerPage={productsPerPage}
+							totalProducts={products.length}
+							paginate={paginate}
+						/>
 					)}
 				</MainBody>
 			)}
